Isoler configImport
