+++ src/components/ImportFromS3Dialog/check-interface-and-sample-type.js
@@ -1,0 +1,31 @@
+import isEmpty from "../../utils/isEmpty"
+export default (typeAuthorize, file) => {
+  function interfaceFileType(type) {
+    if (
+      type === "image_classification" ||
+      type === "image_segmentation" ||
+      type === "composite"
+    )
+      return "Image"
+    if (type === "video_segmentation") return "Video"
+    if (type === "audio_transcription") return "Audio"
+    if (type === "data_entry") return "PDF"
+    if (isEmpty(type)) return "Empty"
+    return "File"
+  }
+
+  function typesamplesSample(samples) {
+    if (isEmpty(samples) || isEmpty(samples[0])) return "Empty"
+    if (!isEmpty(samples[0].imageUrl)) return "Image"
+    if (!isEmpty(samples[0].videoUrl)) return "Video"
+    if (!isEmpty(samples[0].audioUrl)) return "Audio"
+    if (!isEmpty(samples[0].pdfUrl)) return "PDF"
+    return "File"
+  }
+  var result = [null, null]
+  result[0] = interfaceFileType(file.content.interface.type)
+  result[1] = typesamplesSample(file.content.samples)
+  if (typeAuthorize.includes(result[0]) && typeAuthorize.includes(result[1]))
+    return true
+  return false
+}
